i know there's something still to understand

+++ srcs/requirements/nginx/tools/configure.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-mv /tmp/index.html	/var/www/html
-mv /tmp/css			/var/www/html
-mv /tmp/js			/var/www/html
+if [ -f "/tmp/index.html" ];
+then
+	mv /tmp/index.html	/var/www/html
+	mv /tmp/css			/var/www/html
+	mv /tmp/js			/var/www/html
+fi
 
 nginx -g 'daemon off;'
